@@ -10,17 +10,18 @@
         "laravelcollective/html": "self.version"
     },
     "require": {
-        "php": "^7.2 || ^8.0",
-        "illuminate/http": "^6.0 || ^7.0 || ^8.0 || ^9.0 || ^10.0 || ^11.0",
-        "illuminate/routing": "^6.0 || ^7.0 || ^8.0 || ^9.0 || ^10.0 || ^11.0",
-        "illuminate/session": "^6.0 || ^7.0 || ^8.0 || ^9.0 || ^10.0 || ^11.0",
-        "illuminate/support": "^6.0 || ^7.0 || ^8.0 || ^9.0 || ^10.0 || ^11.0",
-        "illuminate/view": "^6.0 || ^7.0 || ^8.0 || ^9.0 || ^10.0 || ^11.0"
+        "php": "^8.2",
+        "illuminate/http": "^11.0",
+        "illuminate/routing": "^11.0",
+        "illuminate/session": "^11.0",
+        "illuminate/support": "^11.0",
+        "illuminate/view": "^11.0"
     },
     "require-dev": {
-        "illuminate/database": "^6.0 || ^7.0 || ^8.0 || ^9.0 || ^10.0 || ^11.0",
-        "mockery/mockery": "~1.0",
-        "phpunit/phpunit": "~8.5 || ^9.5.10"
+        "illuminate/database": "^11.0",
+        "mockery/mockery": "^1.0",
+        "phpunit/phpunit": "^11.0",
+        "symfony/var-dumper": "^7.2"
     },
     "autoload": {
         "psr-4": {
@@ -41,6 +42,6 @@
             }
         }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true
 }
